requirements/base.txt
git+git://github.com/smacker/django-filebrowser-no-grappelli.git#egg=django-filebrowser
git+git://github.com/ITCase-django/django-tinymce-4#egg=django_tinymce_4
Django==1.8.17

# rocket science
numpy==1.11.2
pandas==0.16.2
CVXcanon==0.0.23.3
cvxopt==1.1.8
cvxpy==0.3.4
scikit-learn==0.17
scipy==0.16.0
statsmodels==0.6.1
django-pandas==0.4.1

# caching
django-redis==4.3.0

requests==2.9.1
beautifulsoup4==4.5.1
lxml==3.6.4

diff-match-patch==20121119
django-anymail==0.5
django-authority==0.9
django-ckeditor-updated==4.4.4
django-common-helpers==0.7.0
django-compat==1.0.7
django-cron==0.4.3
celery==4.0.2


django-filter==0.12.0
django-flat-theme==1.1.3
django-haystack==2.4.1
django-import-export==0.4.0
django-localflavor-au-tjh==1.1.2
django-mptt==0.7.4
django_nose==1.4.3
django-nested-admin==3.0.2

django-page-cms==1.8.1
django-phonenumber-field==1.1.0
django-suit==0.2.15
django-taggit==0.17.5
# Using the below git link until drf-extensions==0.3.0 becomes available in pip.
git+git://github.com/chibisov/drf-extensions@0a82b4c57ab95e53c56e5e71cfe71f5a33812a65 # For nested viewsets
ecos==2.0.4
Markdown==2.6.5
pexpect==4.0.1
Pillow==3.0.0
pydenticon==0.2
pinax-eventlog==1.1.1
psycopg2==2.6.1
ptyprocess==0.5
PyPDF2==1.26.0
pysftp==0.2.8
python-dateutil==2.5.3 # For rrule parsing
pytz==2015.7
tzlocal==1.3
scs==1.1.7
six==1.10.0
git+git://github.com/kennethreitz/tablib@ddc4bd30f22e40ae4e3aa9347747f1f96912191a
toolz==0.7.4
ujson==1.34
sqlparse==0.2.1
weasyprint==0.31
jira==1.0.7

# testing
django-test-without-migrations==0.4

# front end
django-bootstrap3==7.0.1

# api
djangorestframework==3.4.6
django-rest-swagger==2.0.5
plaid-python==1.5.0
google-api-python-client>=1.6.2
MicrosoftHealth>=0.1.0
git+https://github.com/maximebf/python-withings.git

# misc
django-model-utils==2.3.1
# re-enable after fixing # django-notifications-hq>=1.0 # upgrade to 1.0 later
django-extensions==1.6.1 # to visualize models

# softlayer swift cloud storage
python-keystoneclient==3.4.0
python-magic==0.4.12
python-swiftclient==3.0.0

#IB API
git+git://github.com/blampe/IbPy.git

# main migrations rely on this
django-recurrence==1.3.0


# geolocation
pygeoip==0.3.2

jira==1.0.7


# ib onboarding
python-gnupg==0.4.0
